backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
aiofiles==23.2.1
httpx==0.27.0
google-generativeai==0.8.6
faiss-cpu==1.8.0
sentence-transformers==2.7.0
xgboost==2.0.3
scikit-learn==1.4.2
joblib==1.4.2
pdfplumber==0.11.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
